backend/python/requirements.txt
Flask
Flask-Cors
rembg
Pillow